Done cart, start transaction

+++ src/main/java/fsse2305/eshop/service/impl/CartItemServiceImpl.java
@@ -5,9 +5,11 @@ import fsse2305.eshop.data.GetCartItemResponseData;
 import fsse2305.eshop.data.PutCartItemResponseData;
 import fsse2305.eshop.data.UpdateCartItemQtyResponseData;
 import fsse2305.eshop.data.entity.CartItemEntity;
+import fsse2305.eshop.data.entity.ProductEntity;
+import fsse2305.eshop.exception.cart.CART_ITEM_EXCEED_STOCK_QTY_EXCEPTION;
 import fsse2305.eshop.exception.cart.CART_ITEM_EXCEPTION;
 import fsse2305.eshop.exception.cart.CART_ITEM_NOT_FOUND_EXCEPTION;
-import fsse2305.eshop.exception.cart.ITEM_ALREADY_ADDED_EXCEPTION;
+import fsse2305.eshop.exception.general.DATA_INSUFFICIENT_EXCEPTION;
 import fsse2305.eshop.repository.CartItemRepository;
 import fsse2305.eshop.service.CartItemService;
 import fsse2305.eshop.service.ProductService;
@@ -35,11 +37,19 @@ public class CartItemServiceImpl implements CartItemService {
 
     public PutCartItemResponseData putCartItem(Integer pid, Integer quantity, FirebaseUserData firebaseUserData) throws Exception {
         try {
-            Integer uid = userService.getEntityByFirebaseUserData(firebaseUserData).getUid();
-            productService.getProductEntityById(pid);
-            if(cartItemRepository.getCartItemByUidAAndPid(uid, pid) != null)    {
-                throw new ITEM_ALREADY_ADDED_EXCEPTION(pid);
+            if(pid == null || quantity == null || firebaseUserData == null) {
+                throw new DATA_INSUFFICIENT_EXCEPTION();
             }
+            Integer uid = userService.getEntityByFirebaseUserData(firebaseUserData).getUid();
+            ProductEntity productEntity = productService.getProductEntityById(pid);
+            if(cartItemRepository.getCartItemByUidAAndPid(uid, pid) != null)    {
+                logger.info("Cart Item already exist, update quantity instead");
+                cartItemRepository.updateCartItemByPid(uid, pid, quantity);
+            }
+            if(productEntity.getStockQty() < quantity)    {
+                throw new CART_ITEM_EXCEED_STOCK_QTY_EXCEPTION(pid, productEntity.getStockQty(), quantity);
+            }
+            logger.info("Add item to cart. uid:" + uid + " pid:" + pid + " Qty:" + quantity);
             Integer result = cartItemRepository.addItem2Cart(uid, pid, quantity);
             if (result==1) {
                 return new PutCartItemResponseData("SUCCESS");
@@ -52,24 +62,36 @@ public class CartItemServiceImpl implements CartItemService {
     }
 
     public List<GetCartItemResponseData> getCartItem(FirebaseUserData firebaseUserData) throws Exception {
+        if(firebaseUserData == null) {
+            throw new DATA_INSUFFICIENT_EXCEPTION();
+        }
         List<GetCartItemResponseData> getCartItemResponseDataList = new ArrayList<>();
-        for(CartItemEntity cartItemEntity: cartItemRepository.getCartItemByUid(userService.getEntityByFirebaseUserData(firebaseUserData).getUid()))   {
+        Integer uid = userService.getEntityByFirebaseUserData(firebaseUserData).getUid();
+        for(CartItemEntity cartItemEntity: cartItemRepository.getCartItemByUid(uid))   {
             getCartItemResponseDataList.add(new GetCartItemResponseData(productService.getProductEntityById(cartItemEntity.getPid()), cartItemEntity));
         }
+        logger.info("Get cart item. uid:" + uid);
         return getCartItemResponseDataList;
     }
 
     public UpdateCartItemQtyResponseData updateCartItemQty(Integer pid, Integer quantity, FirebaseUserData firebaseUserData) throws Exception {
         try {
+            if(pid == null || quantity == null || firebaseUserData == null) {
+                throw new DATA_INSUFFICIENT_EXCEPTION();
+            }
             Integer uid = userService.getEntityByFirebaseUserData(firebaseUserData).getUid();
-            productService.getProductEntityById(pid);
-            if(cartItemRepository.getCartItemByUidAAndPid(uid, pid) == null)    {
+            ProductEntity productEntity = productService.getProductEntityById(pid);
+            CartItemEntity cartItemEntity = cartItemRepository.getCartItemByUidAAndPid(uid, pid);
+            if(cartItemEntity == null)    {
                 throw new CART_ITEM_NOT_FOUND_EXCEPTION(pid);
             }
+            if(productEntity.getStockQty() < quantity)    {
+                throw new CART_ITEM_EXCEED_STOCK_QTY_EXCEPTION(pid, productEntity.getStockQty(), quantity);
+            }
+            logger.info("Update cart item. uid:" + uid + " pid:" + pid + " Qty:" + quantity);
             Integer result = cartItemRepository.updateCartItemByPid(uid, pid, quantity);
             if(result > 0)  {
-                CartItemEntity cartItemEntity = cartItemRepository.getCartItemByUidAAndPid(uid, pid);
-                return new UpdateCartItemQtyResponseData(productService.getProductEntityById(cartItemEntity.getPid()), cartItemEntity);
+                return new UpdateCartItemQtyResponseData(productEntity, quantity);
             }
             throw new CART_ITEM_EXCEPTION();
         }   catch (Exception e) {
@@ -80,11 +102,15 @@ public class CartItemServiceImpl implements CartItemService {
 
     public DeleteCartItemResponseData deleteCartItem(Integer pid, FirebaseUserData firebaseUserData) throws Exception   {
         try {
+            if(pid == null || firebaseUserData == null) {
+                throw new DATA_INSUFFICIENT_EXCEPTION();
+            }
             Integer uid = userService.getEntityByFirebaseUserData(firebaseUserData).getUid();
             productService.getProductEntityById(pid);
             if(cartItemRepository.getCartItemByUidAAndPid(uid, pid) == null)    {
                 throw new CART_ITEM_NOT_FOUND_EXCEPTION(pid);
             }
+            logger.info("Delete cart item. uid:" + uid + " pid:" + pid);
             Integer result = cartItemRepository.deleteCartItemByPid(uid, pid);
             if (result==1) {
                 return new DeleteCartItemResponseData("SUCCESS");
